+++ test/samples/outline.clj
@@ -1,0 +1,19 @@
+(ns outline)
+
+
+;;; First heading level 1
+
+(defn foo
+  [bar]
+  (println bar))
+
+;;;; Heading level 2
+
+(def baz
+  {:hello "World"})
+
+;;; Second heading level 1
+
+(defn hello-world
+  []
+  (println "Hello, world!"))